moved collapsible_lists to components folder
